update the config-set
